Using the new auth `pathTo` does not work with parametric paths.
If you have a route to a path like 'page/:id' and you that requires authentication you will also need to set the pathTo to 'page/:id' but this doesn't parameterize the path it actually just treats it like a static string and send you back to 'page/:id'.

In actuality I would prefer that the redirect to path be the path I just visited. So if I go to 'page/1' I would like to be taken back to 'page/1' after logging in.

Here is a screencast to explain the issue and the fix: http://screencast.com/t/P32218p4